Add temp memory cache for trending items
